refactor(coupon-tag): rename `discount-tag` to `coupon-tag`
